requirements.txt
# Core dependencies
torch>=1.9.0
transformers>=4.12.0
datasets>=1.13.0
librosa>=0.8.1
SoundFile>=0.10.3
pyaudio>=0.2.11
pydub>=0.25.1

# Rasa dependencies
rasa>=3.0.0
rasa-sdk>=3.0.0

# TTS dependencies
gTTS>=2.2.3
pyttsx3>=2.90

# Utility dependencies
numpy>=1.20.0
scipy>=1.7.0
requests>=2.26.0

# Audio processing
soundfile>=0.10.3.post1

# Data handling
pandas>=1.1.0
pyyaml>=5.4.0

# Utilities
tqdm>=4.49.0
pathlib>=1.0.1
python-dateutil>=2.8.1